Move cargo lockfile module under lockfile
