+++ packages/ploys/src/package/cargo/mod.rs
@@ -2,10 +2,8 @@
 
 mod dependency;
 mod error;
-mod lockfile;
 mod manifest;
 
 pub use self::dependency::{Dependencies, DependenciesMut, Dependency, DependencyMut};
 pub use self::error::Error;
-pub use self::lockfile::CargoLockfile;
 pub use self::manifest::CargoManifest;
